Rename all `closest` to `ancestor`
