+++ src/main/java/com/codeborne/selenide/commands/Ancestor.java
@@ -2,7 +2,7 @@ package com.codeborne.selenide.commands;
 
 import com.codeborne.selenide.Command;
 import com.codeborne.selenide.SelenideElement;
-import com.codeborne.selenide.commands.closest.ClosestRuleEngine;
+import com.codeborne.selenide.commands.ancestor.AncestorRuleEngine;
 import com.codeborne.selenide.impl.WebElementSource;
 import org.openqa.selenium.By;
 
@@ -24,7 +24,7 @@ public class Ancestor implements Command<SelenideElement> {
       (args[1] instanceof Integer ? (int) args[1] + 1 : 1) :
       1;
 
-    ClosestRuleEngine ruleEngine = new ClosestRuleEngine();
+    AncestorRuleEngine ruleEngine = new AncestorRuleEngine();
     String xpath = ruleEngine.process(selector, indexPredicate).getValue();
     return locator.find(proxy, By.xpath(xpath), 0);
   }
