+++ src/main/java/com/codeborne/selenide/commands/ancestor/AncestorResult.java
@@ -1,0 +1,14 @@
+package com.codeborne.selenide.commands.ancestor;
+
+public class AncestorResult {
+
+  private final String value;
+
+  public AncestorResult(String value) {
+    this.value = value;
+  }
+
+  public String getValue() {
+    return value;
+  }
+}
