+++ src/main/java/com/codeborne/selenide/commands/ancestor/AncestorRule.java
@@ -1,0 +1,7 @@
+package com.codeborne.selenide.commands.ancestor;
+
+public interface AncestorRule {
+  boolean evaluate(String selector, int index);
+
+  AncestorResult getAncestorResult();
+}
