+++ src/main/java/com/codeborne/selenide/commands/ancestor/AncestorWithAttributeRule.java
@@ -1,8 +1,8 @@
-package com.codeborne.selenide.commands.closest;
+package com.codeborne.selenide.commands.ancestor;
 
 import static java.lang.String.format;
 
-public class ClosestWithAttributeRule extends SelectorValidation implements ClosestRule {
+public class AncestorWithAttributeRule extends SelectorValidation implements AncestorRule {
 
   private String xpath;
 
@@ -18,7 +18,7 @@ public class ClosestWithAttributeRule extends SelectorValidation implements Clos
   }
 
   @Override
-  public ClosestResult getClosestResult() {
-    return new ClosestResult(xpath);
+  public AncestorResult getAncestorResult() {
+    return new AncestorResult(xpath);
   }
 }
