--- src/main/java/com/codeborne/selenide/commands/closest/ClosestResult.java
@@ -1,14 +1,0 @@
-package com.codeborne.selenide.commands.closest;
-
-public class ClosestResult {
-
-  private final String value;
-
-  public ClosestResult(String value) {
-    this.value = value;
-  }
-
-  public String getValue() {
-    return value;
-  }
-}
